Avoid negative characters to isspace

+++ src/options.cpp
@@ -93,7 +93,7 @@ void option_base::write(std::ostream &os, std::string_view section_name, std::si
 		if (std::exchange(do_indent, true))
 			os << indent_str;
 
-		while (not line.empty() and std::isspace(line.back()))
+		while (not line.empty() and std::isspace(static_cast<uint8_t>(line.back())))
 			line.remove_suffix(1);
 
 		os << line << '\n';
